requirements.txt
numpy==1.16.4
networkx==2.3.0
scipy==1.3.1
scikit-learn==0.21.2
gensim==3.6.0
theano==1.0.4
tensorboard==1.15.0
tensorflow==1.15.2
ogb
python-igraph
pymetis
infomap
oslom_runner
